app: add vpn tunneling benchmark with simulated network with different bandwidth/latency
