+++ application_test.go
@@ -1,17 +1,12 @@
 package awl
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"encoding/hex"
-	"errors"
 	"fmt"
 	"io"
 	"net"
 	"net/http"
 	"net/url"
-	"os"
 	"runtime"
 	"strings"
 	"sync"
@@ -19,42 +14,20 @@ import (
 	"testing"
 	"time"
 
-	ds "github.com/ipfs/go-datastore"
-	dssync "github.com/ipfs/go-datastore/sync"
-	"github.com/ipfs/go-log/v2"
-	"github.com/libp2p/go-libp2p"
-	dht "github.com/libp2p/go-libp2p-kad-dht"
-	"github.com/libp2p/go-libp2p/core/peer"
-	"github.com/libp2p/go-libp2p/p2p/host/eventbus"
-	"github.com/libp2p/go-libp2p/p2p/host/peerstore/pstoremem"
-	rcmgr "github.com/libp2p/go-libp2p/p2p/host/resource-manager"
-	"github.com/multiformats/go-multiaddr"
 	"github.com/quic-go/quic-go/integrationtests/tools/israce"
-	"github.com/stretchr/testify/require"
 	"go.uber.org/goleak"
-	"go.uber.org/zap/zapcore"
 	"golang.org/x/net/proxy"
-	"golang.zx2c4.com/wireguard/tun"
 
 	"github.com/anywherelan/awl/api"
-	"github.com/anywherelan/awl/api/apiclient"
 	"github.com/anywherelan/awl/config"
 	"github.com/anywherelan/awl/entity"
-	"github.com/anywherelan/awl/p2p"
-	"github.com/anywherelan/awl/vpn"
 )
-
-func init() {
-	// TODO: move to config
-	useAwldns = false
-	config.DefaultBootstrapPeers = nil
-}
 
 func TestMakeFriends(t *testing.T) {
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
 
 	ts.makeFriends(peer2, peer1)
 }
@@ -62,8 +35,8 @@ func TestMakeFriends(t *testing.T) {
 func TestRemovePeer(t *testing.T) {
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
 
 	ts.makeFriends(peer2, peer1)
 
@@ -112,15 +85,13 @@ func TestRemovePeer(t *testing.T) {
 	// test ping
 	p1Ping := peer1.app.P2p.GetPeerLatency(peer2.app.P2p.PeerID())
 	ts.NotEmpty(p1Ping)
-	p2Ping := peer2.app.P2p.GetPeerLatency(peer1.app.P2p.PeerID())
-	ts.NotEmpty(p2Ping)
 }
 
 func TestDeclinePeerFriendRequest(t *testing.T) {
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
 	ts.ensurePeersAvailableInDHT(peer1, peer2)
 
 	err := peer1.api.SendFriendRequest(peer2.PeerID(), "peer_2")
@@ -149,8 +120,8 @@ func TestDeclinePeerFriendRequest(t *testing.T) {
 func TestAutoAcceptFriendRequest(t *testing.T) {
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
 	ts.ensurePeersAvailableInDHT(peer1, peer2)
 
 	peer2.app.Conf.Lock()
@@ -181,9 +152,9 @@ func TestAutoAcceptFriendRequest(t *testing.T) {
 func TestUniquePeerAlias(t *testing.T) {
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
-	peer3 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
+	peer3 := ts.NewTestPeer(false)
 	ts.ensurePeersAvailableInDHT(peer1, peer2)
 	ts.ensurePeersAvailableInDHT(peer2, peer3)
 
@@ -199,8 +170,8 @@ func TestUniquePeerAlias(t *testing.T) {
 func TestUpdateUseAsExitNodeConfig(t *testing.T) {
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
 
 	ts.makeFriends(peer2, peer1)
 
@@ -281,7 +252,10 @@ func TestUpdateUseAsExitNodeConfig(t *testing.T) {
 		return !peer2Config.AllowedUsingAsExitNode && peer2Config.WeAllowUsingAsExitNode
 	}, 15*time.Second, 100*time.Millisecond)
 
-	ts.Equal("", peer1.app.Conf.SOCKS5.UsingPeerID)
+	peer1.app.Conf.Lock()
+	peer1Socks5UsingPeerID := peer1.app.Conf.SOCKS5.UsingPeerID
+	peer1.app.Conf.Unlock()
+	ts.Equal("", peer1Socks5UsingPeerID)
 
 	availableProxies, err = peer1.api.ListAvailableProxies()
 	ts.NoError(err)
@@ -317,9 +291,9 @@ func TestUpdateUseAsExitNodeConfig(t *testing.T) {
 func TestUpdatePeerSettingsIPAddr(t *testing.T) {
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
-	peer3 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
+	peer3 := ts.NewTestPeer(false)
 
 	// Make peer2 and peer1 friends using the helper
 	ts.makeFriends(peer2, peer1)
@@ -688,8 +662,8 @@ func TestTunnelPackets(t *testing.T) {
 
 	ts := NewTestSuite(t)
 
-	peer1 := ts.newTestPeer(false)
-	peer2 := ts.newTestPeer(false)
+	peer1 := ts.NewTestPeer(false)
+	peer2 := ts.NewTestPeer(false)
 
 	ts.makeFriends(peer2, peer1)
 
@@ -704,7 +678,7 @@ func TestTunnelPackets(t *testing.T) {
 
 	wg := &sync.WaitGroup{}
 
-	sendPackets := func(peer, peerWithInbound testPeer) {
+	sendPackets := func(peer, peerWithInbound TestPeer) {
 		defer wg.Done()
 		packet := testPacket(packetSize)
 
@@ -736,8 +710,8 @@ func BenchmarkTunnelPackets(b *testing.B) {
 		b.Run(fmt.Sprintf("%d bytes per package", packetSize), func(b *testing.B) {
 			ts := NewTestSuite(b)
 
-			peer1 := ts.newTestPeer(true)
-			peer2 := ts.newTestPeer(true)
+			peer1 := ts.NewTestPeer(true)
+			peer2 := ts.NewTestPeer(true)
 
 			ts.makeFriends(peer2, peer1)
 			b.ResetTimer()
@@ -765,308 +739,4 @@ func BenchmarkTunnelPackets(b *testing.B) {
 			b.ReportMetric(packetLoss, "packet_loss")
 		})
 	}
-}
-
-func testPacket(length int) []byte {
-	return testPacketWithDest(length, "10.66.0.2")
-}
-
-func testPacketWithDest(length int, destIP string) []byte {
-	data, err := hex.DecodeString("4500002828f540004011fd490a4200010a420002a9d0238200148bfd68656c6c6f20776f726c6421")
-	if err != nil {
-		panic(err)
-	}
-
-	packet := data
-	if length > len(data) {
-		packet = make([]byte, length)
-		copy(packet, data)
-		_, err = rand.Read(packet[len(data):])
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	vpnPacket := vpn.Packet{}
-	_, err = vpnPacket.ReadFrom(bytes.NewReader(packet))
-	if err != nil {
-		panic(err)
-	}
-	vpnPacket.Parse()
-
-	destIPParsed := net.ParseIP(destIP).To4()
-	if destIPParsed == nil {
-		panic(fmt.Sprintf("invalid destination IP: %s", destIP))
-	}
-	copy(vpnPacket.Dst, destIPParsed)
-
-	vpnPacket.RecalculateChecksum()
-
-	return vpnPacket.Packet
-}
-
-// TODO: add support for goleak in TestSuite
-type TestSuite struct {
-	*require.Assertions
-
-	t                 testing.TB
-	bootstrapAddrs    []peer.AddrInfo
-	bootstrapAddrsStr []string
-}
-
-func NewTestSuite(t testing.TB) *TestSuite {
-	if os.Getenv("CI") != "" && runtime.GOOS == "linux" {
-		t.Skip("doesn't work on linux in CI, flaky ensurePeersAvailableInDHT can't find peers")
-	}
-
-	ts := &TestSuite{t: t, Assertions: require.New(t)}
-	ts.initBootstrapNode()
-	ts.initBootstrapNode()
-
-	return ts
-}
-
-type testPeer struct {
-	app *Application
-	api *apiclient.Client
-	tun *TestTUN
-}
-
-func (tp testPeer) Close() {
-	tp.app.Close()
-}
-
-func (tp testPeer) PeerID() string {
-	return tp.app.Conf.P2pNode.PeerID
-}
-
-func (ts *TestSuite) newTestPeer(disableLogging bool) testPeer {
-	tempDir := ts.t.TempDir()
-	ts.t.Setenv(config.AppDataDirEnvKey, tempDir)
-	tempConf := config.NewConfig(eventbus.NewBus())
-	if disableLogging {
-		tempConf.LoggerLevel = "fatal"
-		log.SetAllLoggers(log.LevelFatal)
-	}
-	tempConf.Save()
-
-	app := New()
-	app.SetupLoggerAndConfig()
-	if disableLogging {
-		log.SetupLogging(zapcore.NewNopCore(), func(string) zapcore.Level {
-			return zapcore.FatalLevel
-		})
-	}
-	app.Conf.HttpListenAddress = "127.0.0.1:0"
-	app.Conf.HttpListenOnAdminHost = false
-	app.Conf.SetListenAddresses([]multiaddr.Multiaddr{
-		multiaddr.StringCast("/ip4/127.0.0.1/tcp/0"),
-		multiaddr.StringCast("/ip4/127.0.0.1/udp/0/quic-v1"),
-	})
-	app.Conf.P2pNode.BootstrapPeers = ts.bootstrapAddrsStr
-	app.Conf.SOCKS5 = config.SOCKS5Config{
-		ListenerEnabled: true,
-		ProxyingEnabled: true,
-		ListenAddress:   pickFreeAddr(ts.t),
-		UsingPeerID:     "",
-	}
-
-	testTUN := NewTestTUN()
-	err := app.Init(context.Background(), testTUN.TUN())
-	ts.NoError(err)
-
-	tp := testPeer{
-		app: app,
-		api: apiclient.New(app.Api.Address()),
-		tun: testTUN,
-	}
-
-	ts.t.Cleanup(func() {
-		tp.Close()
-	})
-
-	return tp
-}
-
-func (ts *TestSuite) initBootstrapNode() {
-	peerstore, err := pstoremem.NewPeerstore()
-	ts.NoError(err)
-	resourceLimitsConfig := rcmgr.InfiniteLimits
-	mgr, err := rcmgr.NewResourceManager(rcmgr.NewFixedLimiter(resourceLimitsConfig))
-	ts.NoError(err)
-
-	hostConfig := p2p.HostConfig{
-		PrivKeyBytes: nil,
-		ListenAddrs: []multiaddr.Multiaddr{
-			multiaddr.StringCast("/ip4/127.0.0.1/tcp/0"),
-			multiaddr.StringCast("/ip4/127.0.0.1/udp/0/quic-v1"),
-		},
-		UserAgent:                config.UserAgent,
-		BootstrapPeers:           ts.bootstrapAddrs,
-		AllowEmptyBootstrapPeers: true,
-		Libp2pOpts: []libp2p.Option{
-			libp2p.DisableRelay(),
-			libp2p.ForceReachabilityPublic(),
-			libp2p.ResourceManager(mgr),
-		},
-		Peerstore:    peerstore,
-		DHTDatastore: dssync.MutexWrap(ds.NewMapDatastore()),
-		DHTOpts: []dht.Option{
-			dht.Mode(dht.ModeServer),
-		},
-	}
-
-	p2pSrv := p2p.NewP2p(context.Background())
-	p2pHost, err := p2pSrv.InitHost(hostConfig)
-	ts.NoError(err)
-	err = p2pSrv.Bootstrap()
-	ts.NoError(err)
-
-	peerInfo := peer.AddrInfo{ID: p2pHost.ID(), Addrs: p2pHost.Addrs()}
-	ts.bootstrapAddrs = append(ts.bootstrapAddrs, peerInfo)
-	addrs, err := peer.AddrInfoToP2pAddrs(&peerInfo)
-	ts.NoError(err)
-	for _, addr := range addrs {
-		ts.bootstrapAddrsStr = append(ts.bootstrapAddrsStr, addr.String())
-	}
-
-	ts.t.Cleanup(func() {
-		_ = p2pSrv.Close()
-	})
-}
-
-func (ts *TestSuite) ensurePeersAvailableInDHT(peer1, peer2 testPeer) {
-	ts.Eventually(func() bool {
-		err1 := peer1.app.P2p.Bootstrap()
-		err2 := peer2.app.P2p.Bootstrap()
-		if err1 != nil || err2 != nil {
-			return false
-		}
-
-		_, err1 = peer1.app.P2p.FindPeer(context.Background(), peer2.app.P2p.PeerID())
-		_, err2 = peer2.app.P2p.FindPeer(context.Background(), peer1.app.P2p.PeerID())
-
-		return err1 == nil && err2 == nil
-	}, 20*time.Second, 100*time.Millisecond)
-}
-
-func (ts *TestSuite) makeFriends(peer1, peer2 testPeer) {
-	ts.ensurePeersAvailableInDHT(peer1, peer2)
-	err := peer1.api.SendFriendRequest(peer2.PeerID(), "peer_2")
-	ts.NoError(err)
-
-	var authRequests []entity.AuthRequest
-	ts.Eventually(func() bool {
-		authRequests, err = peer2.api.AuthRequests()
-		ts.NoError(err)
-		return len(authRequests) == 1
-	}, 15*time.Second, 50*time.Millisecond)
-	err = peer2.api.ReplyFriendRequest(authRequests[0].PeerID, "peer_1", false)
-	ts.NoError(err)
-
-	time.Sleep(500 * time.Millisecond)
-	ts.Len(peer2.app.AuthStatus.GetIngoingAuthRequests(), 0)
-	knownPeer, exists := peer2.app.Conf.GetPeer(peer1.PeerID())
-	ts.True(exists)
-	ts.True(knownPeer.Confirmed)
-
-	knownPeer, exists = peer1.app.Conf.GetPeer(peer2.PeerID())
-	ts.True(exists)
-	ts.True(knownPeer.Confirmed)
-}
-
-type TestTUN struct {
-	Outbound                  chan []byte
-	ReferenceInboundPacketLen int
-
-	inboundCount int64
-	closed       chan struct{}
-	events       chan tun.Event
-	tun          testTun
-}
-
-func NewTestTUN() *TestTUN {
-	c := &TestTUN{
-		Outbound: make(chan []byte),
-		closed:   make(chan struct{}),
-		events:   make(chan tun.Event, 1),
-	}
-	c.tun.t = c
-	c.events <- tun.EventUp
-	return c
-}
-
-func (c *TestTUN) TUN() tun.Device {
-	return &c.tun
-}
-
-func (c *TestTUN) InboundCount() int64 {
-	return atomic.LoadInt64(&c.inboundCount)
-}
-
-func (c *TestTUN) ClearInboundCount() {
-	atomic.StoreInt64(&c.inboundCount, 0)
-}
-
-type testTun struct {
-	t *TestTUN
-}
-
-func (t *testTun) File() *os.File { return nil }
-
-func (t *testTun) Read(bufs [][]byte, sizes []int, offset int) (n int, err error) {
-	for i, buf := range bufs {
-		select {
-		case <-t.t.closed:
-			return n, os.ErrClosed
-		case msg := <-t.t.Outbound:
-			copyN := copy(buf[offset:], msg)
-			sizes[i] = copyN
-			n++
-		}
-	}
-
-	return n, nil
-}
-
-func (t *testTun) Write(bufs [][]byte, offset int) (n int, err error) {
-	for _, buf := range bufs {
-		msg := buf[offset:]
-		if len(msg) != t.t.ReferenceInboundPacketLen {
-			return n, errors.New("packets length mismatch")
-		}
-		select {
-		case <-t.t.closed:
-			return n, os.ErrClosed
-		default:
-		}
-		atomic.AddInt64(&t.t.inboundCount, 1)
-		n++
-	}
-
-	return n, nil
-}
-
-func (t *testTun) BatchSize() int {
-	return 1
-}
-
-func (t *testTun) Flush() error             { return nil }
-func (t *testTun) MTU() (int, error)        { return vpn.InterfaceMTU, nil }
-func (t *testTun) Name() (string, error)    { return "testTun", nil }
-func (t *testTun) Events() <-chan tun.Event { return t.t.events }
-func (t *testTun) Close() error {
-	close(t.t.closed)
-	close(t.t.events)
-	return nil
-}
-
-func pickFreeAddr(t testing.TB) string {
-	l, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer l.Close()
-
-	return l.Addr().String()
 }
